requirements.txt
djangorestframework
psycopg2-binary
python-decouple
Pillow
drf-yasg
django-filter
python-dotenv # скрытие данных
redis # NoSQL БД - брокер сообщений
celery # обработчик задач/сообщений
python-slugify # для генерации slug
djoser # для работы с пользователем
django-cors-headers # для настройки CORS
djangorestframework-simplejwt # для работы с токенами
django-rest-auth
django-allauth
celery
django-redis
django-rest-passwordreset